changed data type for variable mul

+++ 0x0A-argc_argv/3-mul.c
@@ -14,7 +14,7 @@
 #include <stdlib.h>
 int main(int argc, char *argv[])
 {
-	int mul = 1;
+	long int mul = 1;
 	int d;
 
 	if (argc > 1)
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 		{
 			mul *= atoi(argv[d]);
 		}
-		printf("%d\n", mul);
+		printf("%ld\n", mul);
 	}
 	else
 		printf("Error\n");
